requirements.txt
-i https://pypi.org/simple
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
azure-functions==1.2.0
beautifulsoup4==4.8.2
bs4==0.0.1
cached-property==1.5.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
defusedxml==0.6.0
discord.py==1.3.2
discord==1.0.1
dnspython==1.16.0
exchangelib==3.1.1
helpdev==0.6.10
idna==2.9
importlib-metadata==1.5.0
isodate==0.6.0
lxml==4.5.0
multidict==4.7.5
ntlm-auth==1.4.0
numpy==1.18.2
oauthlib==3.1.0
pandas==1.0.3
psutil==5.7.0
pycparser==2.20
pygments==2.6.1
pyodbc==4.0.28
pypika==0.36.0
python-dateutil==2.8.1
pytz==2019.3
pyyaml==5.3.1
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
requests==2.23.0
six==1.14.0
soupsieve==2.0
sqlalchemy==1.3.15
tzlocal==2.0.0
urllib3==1.25.8
websockets==8.1
yarl==1.4.2
zipp==3.1.0
